requirements.txt
Flask==3.0.3
Pillow==10.4.0
python-dotenv==1.0.1
openai==1.46.0
google-generativeai==0.8.3
